Packet injection and more packets
While I was exploring the reason why disconnect hack worked, I found a remote packet injection.
The reason was that part of the sent message was being processed as a packet with an invalid header,
which caused the client to close. Introducing a valid packet allows packet injection, but no nullbytes
can be used.
Also added more packets